Move class definitions out of main
Continuing abstraction support, still some wrinkles to iron out though.
